+++ diesel_cli/tests/print_schema/print_schema_order/sqlite/expected.rs
@@ -39,3 +39,9 @@ table! {
         id -> Nullable<Integer>,
     }
 }
+
+allow_tables_to_appear_in_same_query!(
+    abc,
+    def,
+    ghi,
+);