+++ diesel_cli/tests/print_schema/print_schema_simple/mysql/expected.rs
@@ -25,3 +25,8 @@ table! {
         id -> Integer,
     }
 }
+
+allow_tables_to_appear_in_same_query!(
+    users1,
+    users2,
+);