+++ diesel_cli/tests/print_schema/print_schema_simple_without_docs/mysql/expected.rs
@@ -9,3 +9,8 @@ table! {
         id -> Integer,
     }
 }
+
+allow_tables_to_appear_in_same_query!(
+    users1,
+    users2,
+);